Add more compilation flags
stricter compilation

+++ src/fheroes2/agg/agg.cpp
@@ -395,7 +395,7 @@ void AGG::LoadWAV( int m82, std::vector<u8> & v )
         // create WAV format
         StreamBuf wavHeader( 44 );
         wavHeader.putLE32( 0x46464952 ); // RIFF
-        wavHeader.putLE32( body.size() + 0x24 ); // size
+        wavHeader.putLE32( static_cast<uint32_t>( body.size() ) + 0x24 ); // size
         wavHeader.putLE32( 0x45564157 ); // WAVE
         wavHeader.putLE32( 0x20746D66 ); // FMT
         wavHeader.putLE32( 0x10 ); // size_t
@@ -498,7 +498,7 @@ void AGG::LoadLOOPXXSoundsInternally( const std::vector<int> & vols )
             // new sound
             if ( 0 != vol ) {
             const std::vector<u8> & v = GetWAV( m82 );
-            const int ch = Mixer::Play( &v[0], v.size(), -1, true );
+            const int ch = Mixer::Play( &v[0], static_cast<uint32_t>( v.size() ), -1, true );
 
             if ( 0 <= ch ) {
                 Mixer::Pause( ch );
@@ -544,7 +544,7 @@ void AGG::PlaySoundInternally( const int m82 )
 
     DEBUG_LOG( DBG_ENGINE, DBG_TRACE, M82::GetString( m82 ) );
     const std::vector<u8> & v = AGG::GetWAV( m82 );
-    const int ch = Mixer::Play( &v[0], v.size(), -1, false );
+    const int ch = Mixer::Play( &v[0], static_cast<uint32_t>( v.size() ), -1, false );
     Mixer::Pause( ch );
     Mixer::Volume( ch, Mixer::MaxVolume() * conf.SoundVolume() / 10 );
     Mixer::Resume( ch );
